// src/types/travel.ts
export interface Travel {
  id: string;
  startDate: string;
  endDate: string;
  slots: number;
  destination: string;
  dailyAllowance?: number | null;
  dailyRate?: number | null;
  halfLastDay: boolean;
  volunteers: string[];
  selectedVolunteers?: string[];
  archived: boolean;
  isLocked?: boolean;
}